+++ sgre/testdata/outparam_cross_null/callee.c
@@ -1,0 +1,9 @@
+#include "vsys.h"
+
+extern void *g_buf;
+
+uint32_t get_node(node_t **out)
+{
+    *out = (node_t *)g_buf;
+    return 1;
+}