+++ sgre/testdata/outparam_cross_null/caller.c
@@ -1,0 +1,16 @@
+#include "vsys.h"
+
+void *g_buf;
+
+void caller(void)
+{
+    node_t *n = NULL;
+    get_node(&n);
+    n->val = 1;
+}
+
+void real_null_deref(void)
+{
+    node_t *n = NULL;
+    n->val = 1;
+}